v5.6.0

+++ admin/Model/AddonsModel.php
@@ -1,6 +1,6 @@
 <?php
 /**
-* @version      5.4.1 15.04.2024
+* @version      5.6.0 01.03.2025
 * @author       MAXXmarketing GmbH
 * @package      Jshopping
 * @copyright    Copyright (C) 2010 webdesigner-profi.de. All rights reserved.
@@ -13,12 +13,15 @@ use Joomla\Component\Jshopping\Site\Lib\JSFactory;
 use Joomla\CMS\Factory;
 use Joomla\CMS\Language\Text;
 use Joomla\Component\Jshopping\Site\Helper\Helper;
+use Joomla\Component\Jshopping\Site\Helper\Error as JSError;
 
 defined('_JEXEC') or die();
 
 class AddonsModel extends BaseadminModel{
 
-    function getList($details = 0){
+    protected $nameTable = 'addon';
+
+    public function getList($details = 0){
         $db = Factory::getDBO();
         $jshopConfig = JSFactory::getConfig();
         $query = "SELECT * FROM `#__jshopping_addons`";
@@ -61,33 +64,83 @@ class AddonsModel extends BaseadminModel{
     
     public function save(array $post){
         $row = JSFactory::getTable('addon');
-        $params = $post['params'];
-        if (!is_array($params)){
-            $params = array();
-        }
-        $dispatcher = Factory::getApplication();
-        $dispatcher->triggerEvent('onBeforeSaveAddons', array(&$params, &$post, &$row));
+        $params = $post['params'] ?? [];
+        $app = Factory::getApplication();
+        $app->triggerEvent('onBeforeSaveAddons', array(&$params, &$post, &$row));
         $row->bind($post);
         $row->setParams($params);
         $row->store();
-		$dispatcher->triggerEvent('onAfterSaveAddons', array(&$params, &$post, &$row));
+		$app->triggerEvent('onAfterSaveAddons', array(&$params, &$post, &$row));
         return $row;
     }
+
+    public function saveconfig(array $post){
+        $row = JSFactory::getTable('addon');
+        $config = $post['config'] ?? [];
+        if (!$config['folder_overrides_view']) {
+            unset($config['folder_overrides_view']);
+        }
+        if (!$config['folder_overrides_js']) {
+            unset($config['folder_overrides_js']);
+        }
+        if (!$config['folder_overrides_css']) {
+            unset($config['folder_overrides_css']);
+        }
+        $app = Factory::getApplication();
+        $app->triggerEvent('onBeforeSaveConfigAddons', array(&$config, &$post, &$row));
+        $row->bind($post);
+        $row->setConfig($config);
+        $row->store();
+		$app->triggerEvent('onBeforeSaveConfigAddons', array(&$config, &$post, &$row));
+        return $row;
+    }
+
+    public function deleteList(array $cid, $msg = 1){
+        $res = [];
+		foreach($cid as $id){
+            $this->delete($id, $msg);
+            $res[$id] = true;
+		}
+        return $res;
+    }
     
-    public function delete($id){
+    public function delete($id, $msg = 1){
         $text = '';
-        $dispatcher = Factory::getApplication();
-        $dispatcher->triggerEvent('onBeforeRemoveAddons', array(&$id));
+        $app = Factory::getApplication();
+        $app->triggerEvent('onBeforeRemoveAddons', array(&$id));
+        $adModel = JSFactory::getModel('Addondependencies');
         $row = JSFactory::getTable('addon');
         $row->load($id);
+        $used = $adModel->getList(['alias' => $row->alias]);
+        if (count($used)) {
+            if ($msg){
+                JSError::raiseWarning("", $row->name." - ".Text::_('JSHOP_ADDON_NO_DELETED'));
+            }
+            return 0;
+        }
         if ($row->uninstall){
             include(JPATH_ROOT.$row->uninstall);
         }
+        $alias = $row->alias;
         $row->delete();
-        $dispatcher->triggerEvent('onAfterRemoveAddons', array(&$id, &$text));
-        if ($text){
-            Factory::getApplication()->enqueueMessage($text, 'message');
+        $adModel->deleteByParent($alias);
+        $app->triggerEvent('onAfterRemoveAddons', array(&$id, &$text));
+        if ($msg && $text){
+            $app->enqueueMessage($text, 'message');
         }
+        return 1;
+    }
+
+    public function publish(array $cid, $flag){
+        foreach($cid as $id){
+            $table = $this->getDefaultTable();
+            $table->load($id);
+            if ($flag) {
+                $table->published();
+            } else {
+                $table->unpublished();
+            }
+		}
     }
 
     public function getListWebCategory() {
